Add GRUB2 fallback handling
If the initrd is not able to mount the root file system (e.g. due to missing
or broken hard disk or file system drivers), then revert to the previous initrd
version.
To detect a broken initrd the current health-checker mechanism cannot be used,
as no write access (including setting the default Btrfs subvolume) is possible.
Instead a flag will be set in the GRUB environment block. If the system is
booting then the health-checker of the running system will delete that flag
again. However if the flag is still set when GRUB is starting it seems that the
previous boot wasn't successful. In that case boot a previous known working
snapshot.

The mechanism is backwards compatible, i.e. it will only store the flag if a
snapshot is selected which is aware of the health-checker flag and thus able to
delete it - otherwise we would end up in ininite reboot loops.