Handle booting non-default snapshots
health-checker used to assume that after an update the system would always boot
into the default snapshot.
This assumption is not always true: If a user or the health-checker GRUB part
selected a different snapshot, the results of that snapshot would be taken as
the results for the default snapshot, e.g. by marking the default snapshot as
successfully booted (instead of the current one) or doing a rollback if that
snapshot would fail.

The behaviour was changed to just reboot (in case of failure) and not set the
snapshot state.