Don't rely on t-u's good states in GRUB
With the introduction of the "apply" command in transactional-update the
snapshots don't require a reboot any more to be registered as a "good"
snapshot. Due to [boo#1048088] the GRUB part of health-checker was using
the t-u file as a workaround which cannot be done any more now, so store
the last good snapshot path from a health-checker perspective
independently now.